+++ chalicelib/configer.py
@@ -14,6 +14,11 @@ DEBUG = True if not environ.get('DEBUG') else bool(environ['DEBUG'])
 SNS_TOPIC_FOR_COVID_CASES = "CovidCasesNotifications"
 DDB_CONFIG_TABLE_NAME = 'config'
 DDB_CASES_TABLE_NAME = 'cases'
+DDB_TOTAL_TABLE_NAME = 'total'
+
+# SOURCES
+REGIONS = ['russia']
+MAX_DEEP_DAY = 15
 
 # RUSSIA SOURCE
-RUSSIA_OFFICIAL_URL = 'https://xn--80aesfpebagmfblc0a.xn--p1ai/'
+RUSSIA_OFFICIAL_URL = 'https://xn--80aesfpebagmfblc0a.xn--p1ai/information/'