+++ chalicelib/russia.py
@@ -1,9 +1,10 @@
 from logging import getLogger
 from datetime import datetime, date
 from time import mktime
+from json import loads as json_loads
 
 from requests import get
-from lxml import html
+from bs4 import BeautifulSoup
 
 from chalicelib.models import CovidCase, CovidTotalCases
 from chalicelib.configer import (
@@ -24,35 +25,40 @@ class RussiaOfficialSource(object):
         self.models = []
 
     def request(self, parse=True):
+        """Get and parse data from source. If parse is False, then just
+        check an availability
+        """
         try:
             response = get(f'{self.base_url}{self.path}')
             response.raise_for_status()
             if parse:
-                self.models = self.parse_data(
-                    html.fromstring(response.content))
+                self.models = self.parse_data(response.content)
             return self
         except Exception as e:
             lg.exception(e)
             raise
 
-    def parse_data(self, tree):
-        """
-        """
+    def parse_data(self, content):
+        """Parsing given content"""
+        soup = BeautifulSoup(content, 'lxml')
         model_today, model_total = CovidCase(), CovidTotalCases()
+        datestr = soup.find('h1', {"class": "cv-section__title"}).text
+        data = json_loads(soup.find('cv-stats-virus').attrs[':stats-data'])
+        actual_date = self.__extract_date(datestr)
 
-        base = '//div[@class="d-map__counter"]'
-        date_xpath = '//div[@class="d-map__title"]/h2/span/text()'
-        total_xpath = f'{base}/div[@class="d-map__counter-item"]/h3/text()'
-        new_xpath = f'{base}/div[@class="d-map__counter-item"]/h3/sup/text()'
-
-        actual_date = self.extract_date(tree.xpath(date_xpath))
         model_today.set_actual_date(actual_date)
         model_total.set_actual_date(actual_date)
 
-        case, recovered, death = tree.xpath(new_xpath)
-        model_today.pop(self._i(case), self._i(recovered), self._i(death))
-        case, recovered, death = tree.xpath(total_xpath)
-        model_total.pop(self._i(case), self._i(recovered), self._i(death))
+        model_today.pop(
+            self._i(data['sickChange']),
+            self._i(data['healedChange']),
+            self._i(data['diedChange'])
+        )
+        model_total.pop(
+            self._i(data['sick']),
+            self._i(data['healed']),
+            self._i(data['died'])
+        )
 
         model_today.set_request_datetime()
         model_total.set_request_datetime()
@@ -61,11 +67,15 @@ class RussiaOfficialSource(object):
 
         return model_today.self_validate(), model_total.self_validate()
 
-    def extract_date(self, datelist):
+    def __extract_date(self, datestr):
+        """Extract date from given datestr, it works only for the current
+        source object
+        """
+        lg.debug(f'extracting date from {datestr}')
         month_trans = {'январ': 1, 'февр': 2, 'март': 3, 'апрел': 4, 'май': 5,
                        'мая': 5, 'июн': 6, 'июл': 7, 'август': 8, 'сентябр': 9,
                        'октябр': 10, 'ноябр': 11, 'декабр': 12}
-        _, _, _, day, month, _ = datelist[0].split()
+        _, _, _, _, _, day, month, _ = datestr.split()
         for m, num in month_trans.items():
             if m in month:
                 month = num
@@ -75,4 +85,6 @@ class RussiaOfficialSource(object):
 
     @staticmethod
     def _i(val):
+        """Clean given str val and returns int"""
+        # TODO: place to utils
         return int(val.replace('+', '').replace(' ', ''))