+++ chalicelib/utils.py
@@ -1,14 +1,9 @@
-"""
-utils module
-"""
-
-
 def get_base_url(request):
-    headers = request.headers
-    proto = headers['x-forwarded-proto']
-    host = headers['host']
-    if request.context.get('stage'):
-        stage = request.context['stage']
-    else:
-        stage = ''
-    return f'{proto}://{host}/{stage}'
+    """Helper for getting a current base URL of application. It can provide to
+    suppor a lot of environments without pre-hardcoded stages
+    """
+    # TODO: find a way to get stage and region
+    gapi_id = request.context['apiId']
+    stage = 'api'
+    region = 'eu-north-1'
+    return f'https://{gapi_id}.execute-api.{region}.amazonaws.com/{stage}'