+++ CHANGELOG.md
@@ -1,5 +1,11 @@
+[0.0.4] ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+- added create-aws-schema for simple local and AWS deploying 
+- implemented blueprints for views
+- added new parsed from Russia source
+- fixed bugs with sending and receiving SNS messages
+
 [0.0.1] ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 - added some basic endpoints
 - integrated with telegram api by webhook
 - integrated SNS and DynamoDB
-- implemented a russia official source for grabbing data
+- implemented a russia official source